src/requirements.txt
psycopg2
google-generativeai
python-dotenv